updated put and post request
included how to handle vendor in post request (create a workflow) the API will always set the vendor as null. In the PUT request (update a workflow) it will take in the vendor passed to the workflow and update it. In the body u just have to provide the vendor id